requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-socketio==5.11.2
pyperclip==1.8.2
win10toast==0.9